% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{gh_phase}
\alias{gh_phase}
\title{Pipeline phase header}
\usage{
gh_phase(phase, name)
}
\arguments{
\item{phase}{[EN] Phase number (integer).\cr
[ESP] Numero de fase (entero).}

\item{name}{[EN] Descriptive name of the phase.\cr
[ESP] Nombre descriptivo de la fase.}
}
\value{
[EN] Invisible \code{NULL}.\cr
[ESP] Invisible \code{NULL}.
}
\description{
[EN] Prints a divider line with the name of the current phase.
}
\details{
[ESP] Imprime una linea divisoria con el nombre de la fase actual.
}
\examples{
gh_phase(1, "Data preparation")
gh_phase(2, "Network construction")
}
