% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{gh_progreso_done}
\alias{gh_progreso_done}
\title{Cerrar la barra de progreso}
\usage{
gh_progreso_done(id)
}
\arguments{
\item{id}{ID devuelto por \code{\link[=gh_progreso_init]{gh_progreso_init()}}.}
}
\value{
Invisible \code{NULL}.
}
\description{
Cerrar la barra de progreso
}
\examples{
\dontrun{
pb <- gh_progreso_init(1)
gh_progreso_tick(pb)
gh_progreso_done(pb)
}
}
