% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{gh_progress_done}
\alias{gh_progress_done}
\title{Close the progress bar}
\usage{
gh_progress_done(id)
}
\arguments{
\item{id}{[EN] ID returned by \code{\link[=gh_progress_init]{gh_progress_init()}}.\cr
[ESP] ID devuelto por \code{\link[=gh_progress_init]{gh_progress_init()}}.}
}
\value{
[EN] Invisible \code{NULL}.\cr
[ESP] Invisible \code{NULL}.
}
\description{
[EN] Closes and finalises the progress bar.
}
\details{
[ESP] Cierra y finaliza la barra de progreso.
}
\examples{
\dontrun{
pb <- gh_progress_init(1)
gh_progress_tick(pb)
gh_progress_done(pb)
}
}
