% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{gh_toc}
\alias{gh_toc}
\title{Calcular tiempo transcurrido}
\usage{
gh_toc(t0)
}
\arguments{
\item{t0}{Objeto \code{proc_time} devuelto por \code{\link[=gh_tic]{gh_tic()}}.}
}
\value{
Tiempo transcurrido en segundos (numeric).
}
\description{
Calcular tiempo transcurrido
}
\examples{
t0 <- gh_tic()
elapsed <- gh_toc(t0)
}
