% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{gh_trait_start}
\alias{gh_trait_start}
\title{Print trait processing start}
\usage{
gh_trait_start(trait_id, trait_name, i, n_total)
}
\arguments{
\item{trait_id}{[EN] EFO/MONDO identifier of the trait.\cr
[ESP] Identificador EFO/MONDO del trait.}

\item{trait_name}{[EN] Human-readable trait name.\cr
[ESP] Nombre legible del trait.}

\item{i}{[EN] Current trait index.\cr
[ESP] Indice del trait actual.}

\item{n_total}{[EN] Total number of traits to process.\cr
[ESP] Total de traits a procesar.}
}
\value{
[EN] Invisible \code{NULL}.\cr
[ESP] Invisible \code{NULL}.
}
\description{
[EN] Prints a console line indicating that the pipeline has started
processing a specific trait.
}
\details{
[ESP] Imprime en consola que el pipeline empieza a procesar un trait
concreto.
}
\examples{
gh_trait_start("EFO_0001645", "alzheimer", 12L, 844L)
}
