% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{harmonize_expression}
\alias{harmonize_expression}
\title{Filtrar y normalizar datos de expresion genica}
\usage{
harmonize_expression(config, dt, genes)
}
\arguments{
\item{config}{Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}

\item{dt}{\code{data.table} validado de expresion.}

\item{genes}{\code{data.table} armonizado de genes.}
}
\value{
\code{data.table} filtrado.
}
\description{
Filtra por nivel de expresion minimo (rna_level) y por genes
en lista de referencia.
}
\examples{
\dontrun{
config <- load_config()
genes  <- harmonize_genes(config, validate_genes(config, read_genes(config)))
expr   <- harmonize_expression(
  config,
  validate_expression(config, read_expression(config)),
  genes
)
}
}
