% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{harmonize_genes}
\alias{harmonize_genes}
\title{Filtrar genes protein-coding}
\usage{
harmonize_genes(config, dt)
}
\arguments{
\item{config}{Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}

\item{dt}{\code{data.table} validado de genes.}
}
\value{
\code{data.table} filtrado con genes protein-coding unicos.
}
\description{
Retiene unicamente genes con \code{biotype == "protein_coding"} y
elimina duplicados por \code{id}.
}
\examples{
\dontrun{
config <- load_config()
genes  <- harmonize_genes(config, validate_genes(config, read_genes(config)))
}
}
