% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{read_genes}
\alias{read_genes}
\title{Leer genes protein-coding desde Parquet}
\usage{
read_genes(config)
}
\arguments{
\item{config}{Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}
}
\value{
\code{data.table} con columnas:
\code{id}, \code{approvedSymbol}, \code{biotype}, \code{approvedName}, \code{genomicLocation}.
}
\description{
Lee el dataset de genes de Open Targets Target y devuelve
un subconjunto de columnas esenciales como \code{data.table}.
}
\examples{
\dontrun{
config <- load_config()
genes  <- read_genes(config)
}
}
