% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{read_interactoma}
\alias{read_interactoma}
\title{Leer interactoma proteina-proteina desde Parquet}
\usage{
read_interactoma(config)
}
\arguments{
\item{config}{Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}
}
\value{
\code{data.table} con columnas:
\code{sourceDatabase}, \code{targetA}, \code{intA}, \code{targetB}, \code{intB}, \code{count}, \code{scoring}.
}
\description{
Lee el dataset de interacciones PPI (IntAct, STRING, etc.)
y devuelve las columnas clave como \code{data.table}.
}
\examples{
\dontrun{
config <- load_config()
ppi    <- read_interactoma(config)
}
}
