% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validate.R
\name{validate_diseases}
\alias{validate_diseases}
\title{Validar ontologia de enfermedades}
\usage{
validate_diseases(config, dt)
}
\arguments{
\item{config}{Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}

\item{dt}{\code{data.table} devuelto por \code{\link[=read_diseases]{read_diseases()}}.}
}
\value{
\code{dt} sin modificaciones.
}
\description{
Comprueba columnas y que id y name no tengan NAs.
}
\examples{
\dontrun{
config <- load_config()
dis    <- validate_diseases(config, read_diseases(config))
}
}
