% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validate.R
\name{validate_interactome}
\alias{validate_interactome}
\title{Validate interactome dataset}
\usage{
validate_interactome(config, dt)
}
\arguments{
\item{config}{[EN] Configuration list created by \code{\link[=load_config]{load_config()}}.\cr
[ESP] Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}

\item{dt}{[EN] \code{data.table} returned by \code{\link[=read_interactome]{read_interactome()}}.\cr
[ESP] \code{data.table} devuelto por \code{\link[=read_interactome]{read_interactome()}}.}
}
\value{
[EN] \code{dt} unmodified.\cr
[ESP] \code{dt} sin modificaciones.
}
\description{
[EN] Checks columns, Ensembl pattern in targetA/B and score range.
}
\details{
[ESP] Comprueba columnas, patron Ensembl en targetA/B y rango del score.
}
\examples{
\dontrun{
config <- load_config()
ppi    <- validate_interactome(config, read_interactome(config))
}
}
