% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{gh_tic}
\alias{gh_tic}
\title{Start a stopwatch}
\usage{
gh_tic()
}
\value{
[EN] \code{proc_time} object with the current time.\cr
[ESP] Objeto \code{proc_time} con el tiempo actual.
}
\description{
[EN] Returns the start time for use with \code{\link[=gh_toc]{gh_toc()}}.
}
\details{
[ESP] Devuelve el tiempo de inicio para usarlo con \code{\link[=gh_toc]{gh_toc()}}.
}
\examples{
t0 <- gh_tic()
Sys.sleep(0.1)
gh_toc(t0)
}
