% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{gh_warn}
\alias{gh_warn}
\title{Pipeline warning message}
\usage{
gh_warn(msg)
}
\arguments{
\item{msg}{[EN] Warning text.\cr
[ESP] Texto del aviso.}
}
\value{
[EN] Invisible \code{NULL}.\cr
[ESP] Invisible \code{NULL}.
}
\description{
[EN] Emits a warning alert via cli.
}
\details{
[ESP] Emite una alerta de aviso via cli.
}
\examples{
gh_warn("3 traits discarded for having fewer than 2 genes.")
}
