% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{harmonize_diseases}
\alias{harmonize_diseases}
\title{Filtrar ontologia de enfermedades}
\usage{
harmonize_diseases(config, dt, gwas_comun, gwas_rara)
}
\arguments{
\item{config}{Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}

\item{dt}{\code{data.table} validado de enfermedades.}

\item{gwas_comun}{\code{data.table} armonizado de gwas_comun.}

\item{gwas_rara}{\code{data.table} armonizado de gwas_rara.}
}
\value{
\code{data.table} filtrado de enfermedades.
}
\description{
Retiene solo enfermedades presentes en al menos uno de los
datasets GWAS (comun o raro) para reducir la ontologia
al subconjunto relevante.
}
\examples{
\dontrun{
config <- load_config()
genes  <- harmonize_genes(config, validate_genes(config, read_genes(config)))
gwas_c <- validate_gwas_common(config, read_gwas_common(config))
gwas_r <- validate_gwas_rare(config, read_gwas_rare(config))
gc     <- harmonize_gwas_common(config, gwas_c, genes)
gr     <- harmonize_gwas_rare(config, gwas_r, genes)
dis    <- harmonize_diseases(
  config,
  validate_diseases(config, read_diseases(config)),
  gc, gr
)
}
}
