% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_harmonize.R
\name{harmonize_expression}
\alias{harmonize_expression}
\title{Filter and normalise gene expression data}
\usage{
harmonize_expression(config, dt, genes)
}
\arguments{
\item{config}{[EN] Configuration list created by \code{\link[=load_config]{load_config()}}.\cr
[ESP] Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}

\item{dt}{[EN] Validated expression \code{data.table}.\cr
[ESP] \code{data.table} validado de expresion.}

\item{genes}{[EN] Harmonised gene \code{data.table}.\cr
[ESP] \code{data.table} armonizado de genes.}
}
\value{
[EN] Filtered \code{data.table}.\cr
[ESP] \code{data.table} filtrado.
}
\description{
[EN] Filters by minimum expression level (rna_level) and by genes in the
reference list.
}
\details{
[ESP] Filtra por nivel de expresion minimo (rna_level) y por genes en
lista de referencia.
}
\examples{
\dontrun{
config <- load_config()
genes  <- harmonize_genes(config, validate_genes(config, read_genes(config)))
expr   <- harmonize_expression(
  config,
  validate_expression(config, read_expression(config)),
  genes
)
}
}
