% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validate.R
\name{validate_diseases}
\alias{validate_diseases}
\title{Validate disease ontology}
\usage{
validate_diseases(config, dt)
}
\arguments{
\item{config}{[EN] Configuration list created by \code{\link[=load_config]{load_config()}}.\cr
[ESP] Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}

\item{dt}{[EN] \code{data.table} returned by \code{\link[=read_diseases]{read_diseases()}}.\cr
[ESP] \code{data.table} devuelto por \code{\link[=read_diseases]{read_diseases()}}.}
}
\value{
[EN] \code{dt} unmodified.\cr
[ESP] \code{dt} sin modificaciones.
}
\description{
[EN] Checks columns and that id and name have no NAs.
}
\details{
[ESP] Comprueba columnas y que id y name no tengan NAs.
}
\examples{
\dontrun{
config <- load_config()
dis    <- validate_diseases(config, read_diseases(config))
}
}
