% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validate.R
\name{validate_expression}
\alias{validate_expression}
\title{Validate gene expression dataset}
\usage{
validate_expression(config, dt)
}
\arguments{
\item{config}{[EN] Configuration list created by \code{\link[=load_config]{load_config()}}.\cr
[ESP] Lista de configuracion creada por \code{\link[=load_config]{load_config()}}.}

\item{dt}{[EN] \code{data.table} returned by \code{\link[=read_expression]{read_expression()}}.\cr
[ESP] \code{data.table} devuelto por \code{\link[=read_expression]{read_expression()}}.}
}
\value{
[EN] \code{dt} unmodified.\cr
[ESP] \code{dt} sin modificaciones.
}
\description{
[EN] Checks columns, Ensembl pattern, rna_value values and rna_level.
}
\details{
[ESP] Comprueba columnas, patron Ensembl, valores de rna_value y
rna_level.
}
\examples{
\dontrun{
config <- load_config()
expr   <- validate_expression(config, read_expression(config))
}
}
